// frontend/src/services/generated/models/CoachNoteResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Fitlytics API
 * Workout tracking and training analytics API. Protected routes require a WorkOS access token in the `Authorization: Bearer ...` header.
 *
 * The version of the OpenAPI document: 0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CoachNoteResponse
 */
export interface CoachNoteResponse {
    /**
     * 
     * @type {string}
     * @memberof CoachNoteResponse
     */
    authorName?: string;
    /**
     * 
     * @type {string}
     * @memberof CoachNoteResponse
     */
    authorUserId?: string;
    /**
     * 
     * @type {string}
     * @memberof CoachNoteResponse
     */
    body?: string;
    /**
     * 
     * @type {string}
     * @memberof CoachNoteResponse
     */
    createdAt?: string;
    /**
     * 
     * @type {string}
     * @memberof CoachNoteResponse
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof CoachNoteResponse
     */
    setVideoId?: string;
}

/**
 * Check if a given object implements the CoachNoteResponse interface.
 */
export function instanceOfCoachNoteResponse(value: object): value is CoachNoteResponse {
    return true;
}

export function CoachNoteResponseFromJSON(json: any): CoachNoteResponse {
    return CoachNoteResponseFromJSONTyped(json, false);
}

export function CoachNoteResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): CoachNoteResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'authorName': json['author_name'] == null ? undefined : json['author_name'],
        'authorUserId': json['author_user_id'] == null ? undefined : json['author_user_id'],
        'body': json['body'] == null ? undefined : json['body'],
        'createdAt': json['created_at'] == null ? undefined : json['created_at'],
        'id': json['id'] == null ? undefined : json['id'],
        'setVideoId': json['set_video_id'] == null ? undefined : json['set_video_id'],
    };
}

export function CoachNoteResponseToJSON(json: any): CoachNoteResponse {
    return CoachNoteResponseToJSONTyped(json, false);
}

export function CoachNoteResponseToJSONTyped(value?: CoachNoteResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'author_name': value['authorName'],
        'author_user_id': value['authorUserId'],
        'body': value['body'],
        'created_at': value['createdAt'],
        'id': value['id'],
        'set_video_id': value['setVideoId'],
    };
}